[Property,Section] Move outside pycore/
