+++ nixio/file.py
@@ -22,7 +22,7 @@ import h5py
 from .pycore.h5group import H5Group
 from .block import Block
 from .exceptions import exceptions
-from .pycore.section import Section
+from .section import Section
 from . import util
