--- nixio/pycore/__init__.py
@@ -1,2 +1,0 @@
-from .property import Property
-from .section import Section
