+++ nixio/pycore/h5group.py
@@ -6,14 +6,13 @@
 # modification, are permitted under the terms of the BSD License. See
 # LICENSE file in the root of the Project.
 from __future__ import (absolute_import, division, print_function)
-
 import h5py
 import numpy as np
 
 from .h5dataset import H5DataSet
 from ..value import DataType
 from ..block import Block
-from .section import Section
+from ..section import Section
 
 from .. import util
 from ..exceptions import InvalidEntity
